+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # aGrUM Changelog
 
+## Changelog for 1.12.0
+
+* aGrUM
+  * add and complete the new class `gum::MeekRules` to complete a learned PDAG to CPADG or BN.
+  * Preventing 'infinities' in the definition of numerical discrete variable.
+
+* pyAgrum
+  * Preventing 'infinities' in the definition of numerical discrete variable.
+  * add a new `expert` discretization method to `pyAgrum.skbn.BNDiscretizer`.
+  * add `pyAgrum.skbn.BNDiscretizer.auditCSV` to directly audit a csv file.
+  * add `pyAgrum.fastGraph` to create `pyAgrum.{Di|Undi|Mixed}Graph`.
+  * Dataframe can now come from `pandas` or `polars` (tahnks to Mads Linksou).
+  * pickle for graphicals models !
+
 ## Changelog for 1.11.0
 
 * Internal